+++ DU-Industry-Tool/Program.cs
@@ -1,4 +1,5 @@
 ﻿using System;
+using System.Diagnostics;
 using System.Threading;
 using System.Windows.Forms;
 
@@ -25,11 +26,27 @@ namespace DU_Industry_Tool
 
         private static void App_UIThreadException(object sender, ThreadExceptionEventArgs e)
         {
-            MessageBox.Show(e.Exception.Message, "UI Thread Exception", MessageBoxButtons.OK, MessageBoxIcon.Error);
+            if (e.Exception != null)
+            {
+                Debug.WriteLine(e.Exception.Message);
+                if (e.Exception.StackTrace != null)
+                {
+                    Debug.WriteLine(e.Exception.StackTrace);
+                    //MessageBox.Show(e.Exception?.Message + "\r\n" + e.Exception.StackTrace,
+                    //    "UI Thread Exception", MessageBoxButtons.OK, MessageBoxIcon.Error);
+                }
+            }
         }
+
         private static void CurrentDomain_UnhandledException(object sender, UnhandledExceptionEventArgs e)
         {
-            MessageBox.Show(((Exception)e.ExceptionObject).Message, "Non-UI Thread Exception", MessageBoxButtons.OK, MessageBoxIcon.Error);
+            if (!(e.ExceptionObject is Exception ex)) return;
+            Debug.WriteLine(ex.Message);
+            if (ex.StackTrace != null)
+            {
+                Debug.WriteLine(ex.StackTrace);
+            }
+            //MessageBox.Show(((Exception)e.ExceptionObject).Message, "Non-UI Thread Exception", MessageBoxButtons.OK, MessageBoxIcon.Error);
         }
     }
 }